Refactor ManagePage components
