--- client/src/components/ManagePageContainer/ManagePageComponents.module.scss
@@ -1,342 +1,0 @@
-@import '../../assets/styles/colors.scss';
-@import '../../assets/styles/mediaQueriesMixings.scss';
-@import '../../assets/styles/fontMixins.scss';
-
-.container {
-  .tableHeader {
-    @include sf-pro-display-medium;
-    font-size: 0.875em;
-  }
-  .tableHeader {
-    position: sticky;
-    top: 0;
-    background-color: $white;
-    padding-bottom: 15px;
-  }
-  .firstHeaderCol {
-    padding-left: 15px;
-  }
-
-  .tableCol {
-    white-space: nowrap;
-    background-color: $milk;
-    margin-top: 10px;
-    margin-bottom: 10px;
-    vertical-align: middle;
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    padding: 10px;
-    font-size: 0.8125em;
-  }
-
-  .firstCol {
-    border-top-left-radius: 15px; 
-    border-bottom-left-radius: 15px;
-    padding-left: 15px;
-    padding-right: 20px;
-    @include sf-pro-display-regular;
-  }
-
-  .lastCol {
-    border-bottom-right-radius: 15px; 
-    border-top-right-radius: 15px;
-    padding-right: 10px;
-  }
-
-  .dataItemContainer {
-    background-color: $milk;
-    border-radius: 15px;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding: 15px;
-    margin-bottom: 20px;
-    font-size: 0.875em;
-  }
-
-  .nameContainer {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-
-    .name {
-      white-space: pre-wrap;
-      margin-bottom: 0;
-      word-break: break-all;
-    }
-  }
-  .openIcon {
-    margin-left: 10px;
-  }
-
-  .remaining {
-    display: flex;
-    margin: 0 auto;
-  }
-}
-
-.domainType {
-  color: $white;
-  background-color: $fountain-blue;
-  padding: 5px;
-  border-radius: 3px;
-  font-size: 0.825em;
-  line-height: 0.825em;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-left: 25px;
-
-  &.public {
-    background-color: $irish-green;
-  }
-
-  @include lg {
-    margin-left: 0;
-    font-size: 0.77em;
-  }
-}
-
-.actionButtonsContainer {
-  display: flex;
-
-  .actionButton {
-    @include sf-pro-display-thin;
-    background-color: $violet;
-    border-color: $violet;
-    border-radius: 7px;
-    margin-right: 20px;
-    white-space: nowrap;
-
-    &:hover {
-      background-color: $clear-blue;
-      border-color: $clear-blue;
-    }
-    
-    & button {
-      font-size: 1em;
-      width: 100%;
-      height: 100%;
-    }
-
-    @include sm {
-      width: 100%;
-      height: 45px !important;
-      display: flex;
-      align-items: center;
-      img, svg {
-        margin-right: 10px;
-      }
-    }
-  }
-
-  .settingsButton {
-    background-color: $manatee;
-    border-color: $manatee;
-    width: 36px;
-    font-size: 1.25em;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    &:hover {
-      background-color: $dove-gray;
-      border-color: $dove-gray;
-    }
-
-    @include sm {
-      height: 45px !important;
-      width: 45px !important
-    }
-  }
-
-  .actionButton, .settingsButton {
-    height: 36px;
-    border-radius: 7px;
-    transition: color .15s ease-in-out,
-    background-color .15s ease-in-out,
-    border-color .15s ease-in-out,
-    box-shadow .15s ease-in-out;
-  }
-
-  .atIcon {
-    margin-right: 10px;
-  }
-}
-
-.infoIcon {
-  margin-left: 10px;
-  color: $blue;
-  cursor: pointer;
-}
-.warnIcon {
-  margin-left: 10px;
-  color: $simple-orange;
-  cursor: pointer;
-}
-
-.itemContainer {
-  .title {
-    font-size: 1.25em;
-    color: $gray-main;
-    margin-left: 20px;
-  }
-
-  .itemNameContainer {
-    margin-top: 30px;
-    display: flex;
-    flex-direction: row;
-
-    .itemName {
-      font-size: 1em;
-      white-space: pre-wrap;
-      word-break: break-all;
-      margin-bottom: 0;
-      padding-left: 20px;
-    }
-  }
-  .domainStatus {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-
-    .domainStatusBadge {
-      margin-left: 20px;
-    }
-  }
-
-  .badgeTitle, .badgeItem {
-    font-size: 0.875em;
-    @include proxima-nova-semibold;
-  }
-  .badgeTitle {
-    color: $manatee;
-    margin: 0;
-  }
-  .badgeItem {
-    color: $black;
-  }
-
-  .itemActions {
-    margin-top: 30px;
-    .actionsTitle {
-      font-size: 1em;
-      margin-left: 20px;
-      margin-bottom: 15px;
-      @include proxima-nova-semibold;
-    }
-  }
-}
-
-.settingsContainer {
-  width: 100%;
-  .title {
-    font-size: 1.875em;
-    @include proxima-nova-semibold;
-  }
-  .subtitle, .actionTitle {
-    font-size: 1em;
-    @include proxima-nova-semibold;
-    margin-top: 1.875em;
-  }
-
-  .subtitle {
-    margin-bottom: -15px;
-  }
-
-  .badgeContainer {
-    display: flex;
-    flex-direction: row;
-  }
-
-  .badgeTitle, .badgeItem {
-    color: $black;
-    @include proxima-nova-semibold;
-    font-size: 0.875em;
-    margin-bottom: 0;
-  }
-
-  .badgeTitle {
-    margin-left: 0;
-    margin-right: 0;
-    white-space: nowrap;
-  }
-
-  .badgeItem {
-    margin-left: 15px;
-    word-break: break-all;
-  }
-
-  .text {
-    font-size: 0.875em;
-    @include proxima-nova-thin;
-  }
-
-  .buttonLink {
-    width: fit-content;
-    display: block;
-
-    .button {
-      display: flex;
-      align-items: center;
-      height: 54px;
-      padding: 20px;
-      border-radius: 7px;
-      font-size: 0.875em;
-      margin-bottom: 20px;
-      @include sf-pro-display-medium;
-    }
-  }
-}
-
-@include sm {
-  .settingsContainer {
-    .title {
-      font-size: 1.25em;
-    }
-
-    .badgeContainer {
-      flex-direction: column;
-
-      .badgeItem {
-        margin-left: 0;
-        margin-top: 1rem;
-      }
-    }
-
-    .buttonLink {
-      .button {
-        height: 36px;
-      }
-    }
-  }
-}
-
-@include lg {
-  .container {
-    display: grid;
-    grid-template-columns: auto auto auto auto;
-    width: fit-content;
-
-    .tableHeader, .tableCol {
-      padding: 10px;
-    }
-    .tableHeader {
-      font-size: 0.8125em;
-    }
-
-    .nameContainer {
-      width: 100%;
-      justify-content: space-between;
-    }
-
-    .warnIcon {
-      margin-left: 5px;
-    }
-  }
-  .settingsContainer {
-    padding: 0 10px;
-  }
-}
