+++ client/src/components/ManagePageContainer/ManagePageComponents/DateComponent.module.scss
@@ -1,0 +1,12 @@
+@import '../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../assets/styles/colors.scss';
+
+.warnIcon {
+  margin-left: 10px;
+  color: $simple-orange;
+  cursor: pointer;
+  
+  @include lg {
+    margin-left: 5px;
+  }
+}
