+++ client/src/components/ManagePageContainer/ManagePageComponents/FioAddress.module.scss
@@ -1,0 +1,20 @@
+@import '../../../assets/styles/colors.scss';
+
+.nameContainer {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+
+  .name {
+    white-space: pre-wrap;
+    margin-bottom: 0;
+    word-break: break-all;
+  }
+}
+
+.infoIcon {
+  margin-left: 10px;
+  color: $blue;
+  cursor: pointer;
+}
+
