+++ client/src/components/ManagePageContainer/ManagePageComponents/ItemComponent.module.scss
@@ -1,0 +1,59 @@
+@import '../../../assets/styles/colors.scss';
+@import '../../../assets/styles/fontMixins.scss';
+
+.title {
+  font-size: 1.25em;
+  color: $gray-main;
+  margin-left: 20px;
+}
+
+.itemNameContainer {
+  margin-top: 30px;
+  display: flex;
+  flex-direction: row;
+
+  .itemName {
+    font-size: 1em;
+    white-space: pre-wrap;
+    word-break: break-all;
+    margin-bottom: 0;
+    padding-left: 20px;
+  }
+
+  .infoIcon {
+    margin-left: 10px;
+    color: $blue;
+    cursor: pointer;
+  }
+}
+.domainStatus {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+
+  .domainStatusBadge {
+    margin-left: 20px;
+  }
+}
+
+.badgeTitle, .badgeItem {
+  font-size: 0.875em;
+  @include proxima-nova-semibold;
+}
+.badgeTitle {
+  color: $manatee;
+  margin: 0;
+}
+.badgeItem {
+  color: $black;
+}
+
+.itemActions {
+  margin-top: 30px;
+  .actionsTitle {
+    font-size: 1em;
+    margin-left: 20px;
+    margin-bottom: 15px;
+    @include proxima-nova-semibold;
+  }
+}
