+++ client/src/components/ManagePageContainer/ManagePageComponents/SettingsItem.module.scss
@@ -1,0 +1,93 @@
+@import '../../../assets/styles/colors.scss';
+@import '../../../assets/styles/fontMixins.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
+
+.settingsContainer {
+  width: 100%;
+  .title {
+    font-size: 1.875em;
+    @include proxima-nova-semibold;
+  }
+  .subtitle, .actionTitle {
+    font-size: 1em;
+    @include proxima-nova-semibold;
+    margin-top: 1.875em;
+  }
+
+  .subtitle {
+    margin-bottom: -15px;
+  }
+
+  .badgeContainer {
+    display: flex;
+    flex-direction: row;
+  }
+
+  .badgeTitle, .badgeItem {
+    color: $black;
+    @include proxima-nova-semibold;
+    font-size: 0.875em;
+    margin-bottom: 0;
+  }
+
+  .badgeTitle {
+    margin-left: 0;
+    margin-right: 0;
+    white-space: nowrap;
+  }
+
+  .badgeItem {
+    margin-left: 15px;
+    word-break: break-all;
+  }
+
+  .text {
+    font-size: 0.875em;
+    @include proxima-nova-thin;
+  }
+
+  .buttonLink {
+    width: fit-content;
+    display: block;
+
+    .button {
+      display: flex;
+      align-items: center;
+      height: 54px;
+      padding: 20px;
+      border-radius: 7px;
+      font-size: 0.875em;
+      margin-bottom: 20px;
+      @include sf-pro-display-medium;
+    }
+  }
+}
+
+@include sm {
+  .settingsContainer {
+    .title {
+      font-size: 1.25em;
+    }
+
+    .badgeContainer {
+      flex-direction: column;
+
+      .badgeItem {
+        margin-left: 0;
+        margin-top: 1rem;
+      }
+    }
+
+    .buttonLink {
+      .button {
+        height: 36px;
+      }
+    }
+  }
+}
+
+@include lg {
+   .settingsContainer {
+    padding: 0 10px;
+  }
+}
