+++ client/src/components/ManagePageContainer/ManagePageComponents/UIView.module.scss
@@ -1,0 +1,98 @@
+@import '../../../assets/styles/colors.scss';
+@import '../../../assets/styles/fontMixins.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
+
+.container {
+  .tableHeader {
+    @include sf-pro-display-medium;
+    font-size: 0.875em;
+  }
+  .tableHeader {
+    position: sticky;
+    top: 0;
+    background-color: $white;
+    padding-bottom: 15px;
+  }
+  .firstHeaderCol {
+    padding-left: 15px;
+  }
+
+  .tableCol {
+    white-space: nowrap;
+    background-color: $milk;
+    margin-top: 10px;
+    margin-bottom: 10px;
+    vertical-align: middle;
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    padding: 10px;
+    font-size: 0.8125em;
+  }
+
+  .firstCol {
+    border-top-left-radius: 15px; 
+    border-bottom-left-radius: 15px;
+    padding-left: 15px;
+    padding-right: 20px;
+    @include sf-pro-display-regular;
+  }
+
+  .lastCol {
+    border-bottom-right-radius: 15px; 
+    border-top-right-radius: 15px;
+    padding-right: 10px;
+  }
+
+  .dataItemContainer {
+    background-color: $milk;
+    border-radius: 15px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    padding: 15px;
+    margin-bottom: 20px;
+    font-size: 0.875em;
+  }
+
+  .nameContainer {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+
+    .name {
+      white-space: pre-wrap;
+      margin-bottom: 0;
+      word-break: break-all;
+    }
+  }
+  .openIcon {
+    margin-left: 10px;
+  }
+
+  .remaining {
+    display: flex;
+    margin: 0 auto;
+  }
+}
+
+@include lg {
+  .container {
+    display: grid;
+    grid-template-columns: auto auto auto auto;
+    width: fit-content;
+
+    .tableHeader, .tableCol {
+      padding: 10px;
+    }
+    .tableHeader {
+      font-size: 0.8125em;
+    }
+
+    .nameContainer {
+      width: 100%;
+      justify-content: space-between;
+    }
+  }
+}
